requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
scikit-image>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
geopy>=2.4.0
Pillow>=10.0.0
opencv-python>=4.8.0
tensorflow>=2.3.1
keras>=2.4.3
streamlit>=1.27.1
matplotlib>=3.3.2
